requirement.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
tensorflow==2.15.0
pillow==10.3.0
numpy==1.26.4
tf-keras
accelerate

